move logic to util
